Refactor NodeViews and ImageViewer. Added Pixel information to ImageRenderer
