+++ nodestudiojs/src/components/ImageViewer/Image3dViewer.scss
@@ -1,3 +1,10 @@
+.image-3d-viwer {
+    h2 {
+        color: #EEEEEE;
+        font-weight: 300;
+    }
+}
+
 .viewer-3d {
     position: relative;
 }
