+++ nodestudiojs/src/components/ImageViewer/ImageMultiLayerRenderer.jsx
