+++ nodestudiojs/src/components/ImageViewer/ImageSimpleRenderer.jsx
@@ -1,6 +1,6 @@
-import { useEffect,useContext, useState } from 'react';
+import { useEffect, useContext, useState, useRef } from 'react';
 import APIDataService from '../../services/APIDataService';
-import { DrawImg } from '../../libraries/draw/Draw';
+import { DrawImg, getImgPixelValue } from '../../libraries/draw/Draw';
 import { decodeDataset } from '../../libraries/signal/Dataset';
 import { throttle } from '../../libraries/utils';
 import { ActionTypes } from '../../state/AppReducers';
@@ -9,8 +9,11 @@ import DefaultImg from  '../../images/default_image_icon.png';
 
 const ImageRender = ({ slice, index, colormap, nodeID }) => {
     const {state, dispatch} = useContext(AppState.AppContext);
+    const imgRef = useRef(null);
     const [imageData, setImageData] = useState(DefaultImg);
-    
+    const [dataset, setDataset] = useState(null);
+    const [position, setPosition] = useState({ x:0, y:0 });
+
     useEffect(() => {
         fetchData(slice, index, colormap);
     // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -27,13 +30,36 @@ const ImageRender = ({ slice, index, colormap, nodeID }) => {
                 const dataset = decodeDataset(encodedData);
                 const dataUri = DrawImg(dataset, colormap);
                 setImageData(dataUri);
+                setDataset(dataset);
                 dispatch({type:ActionTypes.UPDATE_SESSION, nodeID, update:false});
             }
         }, 100, slice);
     }
     
+    const handleMouseMove = (e) => {
+        if(dataset === null) return;
+
+        var rect = imgRef.current.getBoundingClientRect();
+        const width = imgRef.current.clientWidth
+        const height = imgRef.current.clientHeight;
+        const x = Math.round(dataset.shape[1] * (e.pageX - rect.left) / width);
+        const y = Math.round(dataset.shape[0] * (e.pageY - rect.top) / height);
+        setPosition({ x, y });
+    }
+
+    const getImageValue = () => {
+        if(dataset === null) return;
+        return getImgPixelValue(dataset, position.x, position.y)
+    }
+
     return (
-        <img src={imageData} alt='viewport' style = {{width : '100%', height : '100%'}}></img>
+        <div style = {{ width : '100%', height : '100%', position: 'relative' }}>
+            <div style={{ color:'#AAAAAA', margin:'0.5em', position: 'absolute', zIndex:1, fontSize:'0.8em' }}>
+                <div> x: { position.x }, y:{ position.y } </div>
+                <div> value: { getImageValue() } </div>
+            </div>
+            <img src={imageData} alt='viewport' ref={imgRef}  style = {{ width : '100%', height : '100%' }} onMouseMove={handleMouseMove}></img>
+        </div>
     )
 
 }
