+++ nodestudiojs/src/components/ImageViewer/ImageViewer.scss
@@ -2,8 +2,4 @@
     .image-viewer-options {
         margin-bottom: 1em;
     }
-
-    .image-viewer-viewport {
-
-    }
 }
